pyproject.toml
[tool.poetry]
name = "nonebot-plugin-fakemsg"
version = "0.1.2"
description = "适用于onebot的伪消息插件"
authors = ["Cvandia <106718176+Cvandia@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "nonebot_plugin_fakemsg" }]

[tool.poetry.dependencies]
nonebot-adapter-onebot = ">=2.0.0-beta.1"
python = ">=3.8,<4"
nonebot2 = ">=2.0.0-beta.1"
pydantic = ">=1.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
